Preparation for dx12 merge (it's not done, but I'm waiting until shader model 7 so I can use spirv across all backends

+++ src/slag/backends/vulkan/core/IVulkanCommandBuffer.h
@@ -20,7 +20,8 @@ namespace slag
             virtual void end() override;
 
 #ifndef SLAG_DISCREET_TEXTURE_LAYOUTS
-            virtual void insertBarriers(TextureBarrier* textureBarriers, size_t textureBarrierCount,BufferBarrier* bufferBarriers, size_t bufferBarrierCount,GlobalBarrier* memoryBarriers, size_t memoryBarrierCount) override;
+            virtual void insertBarriers(TextureBarrier* textureBarriers, uint32_t textureBarrierCount, BufferBarrier* bufferBarriers, uint32_t
+                                        bufferBarrierCount, GlobalBarrier* memoryBarriers, uint32_t memoryBarrierCount) override;
             virtual void insertBarrier(const TextureBarrier& barrier) override;
             virtual void clearTexture(Texture* texture, ClearColor color) override;
             virtual void clearTexture(Texture* texture, ClearDepthStencilValue depthStencil) override;
@@ -30,7 +31,7 @@ namespace slag
             virtual void blit(Texture* source, uint32_t sourceLayer, uint32_t sourceMip, Rectangle sourceArea,Texture* destination, uint32_t destinationLayer, uint32_t destinationMip,Rectangle destinationArea) override;
             virtual void blit(Texture* source, uint32_t sourceLayer, uint32_t sourceMip, Rectangle sourceArea,Texture* destination, uint32_t destinationLayer, uint32_t destinationMip,Rectangle destinationArea, Pixels::AspectFlags aspects) override;
             virtual void resolve(Texture* source, uint32_t sourceLayer, uint32_t sourceMip, Offset sourceOffset, Texture* destination, uint32_t destinationLayer, uint32_t destinationMip, Offset destinationOffset, Extent resolveExtent) override;
-            virtual void beginRendering(Attachment* colorAttachments, size_t colorAttachmentCount,Attachment* depthAttachment, Rectangle bounds) override;
+            virtual void beginRendering(Attachment* colorAttachments, uint32_t colorAttachmentCount, Attachment* depthAttachment, Rectangle bounds) override;
 
 #else
             virtual void insertBarriers(TextureBarrierDiscreet* textureBarriers, size_t textureBarrierCount, BufferBarrier* bufferBarriers, size_t bufferBarrierCount, GPUMemoryBarrier* memoryBarriers, size_t memoryBarrierCount)override;
@@ -50,7 +51,7 @@ namespace slag
 
             virtual void copyBufferToBuffer(Buffer* source, uint64_t sourceOffset, uint64_t length, Buffer* destination,uint64_t destinationOffset) override;
 
-            virtual void fillBuffer(Buffer* buffer, size_t offset, size_t length, uint32_t data) override;
+            virtual void fillBuffer(Buffer* buffer, uint64_t offset, uint64_t length, uint32_t data) override;
             virtual void setViewPort(float x, float y, float width, float height, float minDepth, float maxDepth) override;
             virtual void setScissors(Rectangle rectangle) override;
             virtual void setBlendConstants(float r, float g, float b, float a) override;
@@ -64,13 +65,13 @@ namespace slag
             virtual void drawIndirectCount(Buffer* buffer, uint64_t offset, Buffer* countBuffer,uint64_t countBufferOffset, uint32_t maxDrawCount, uint32_t stride) override;
             virtual void dispatch(uint32_t groupCountX, uint32_t groupCountY, uint32_t groupCountZ) override;
             virtual void dispatchBase(uint32_t baseGroupX, uint32_t baseGroupY, uint32_t baseGroupZ,uint32_t groupCountX, uint32_t groupCountY, uint32_t groupCountZ) override;
-            virtual void dispatchIndirect(Buffer* buffer, size_t offset) override;
+            virtual void dispatchIndirect(Buffer* buffer, uint64_t offset) override;
             virtual void bindGraphicsShaderPipeline(ShaderPipeline* pipeline) override;
             virtual void bindComputeShaderPipeline(ShaderPipeline* pipeline) override;
             virtual void bindGraphicsDescriptorBundle(uint32_t index, DescriptorBundle& bundle) override;
             virtual void bindComputeDescriptorBundle(uint32_t index, DescriptorBundle& bundle) override;
             virtual void bindIndexBuffer(Buffer* buffer, Buffer::IndexSize indexSize, uint64_t offset) override;
-            virtual void bindVertexBuffers(uint32_t firstBindingIndex, Buffer** buffers, uint64_t* bufferOffsets,uint64_t* strides,size_t bufferCount) override;
+            virtual void bindVertexBuffers(uint32_t firstBindingIndex, Buffer** buffers, uint64_t* bufferOffsets, uint64_t* strides, uint32_t bufferCount) override;
 
 
             void transitionToLayout(Texture* texture,VkImageLayout oldLayout, VkImageLayout newLayout,VkAccessFlags2 accessBefore, VkAccessFlags2 accessAfter, VkPipelineStageFlags2 syncBefore, VkPipelineStageFlags2 syncAfter) const;
